Separate generator logic.
