+++ src/cosi_db/mod.rs
@@ -1,3 +1,4 @@
 pub mod common;
 pub mod connection;
+pub mod generator;
 pub mod person;
